test(HttpRequestConverter): HttpRequest로 변화시키는지 확인하는 테스트 작성
